requirement.txt
certifi==2024.8.30
cffi==1.17.1
circuitbreaker==2.0.0
cryptography==42.0.8
mcrcon==0.7.0
oci==2.133.0
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0